Small refactor, get seekbars to be at the right position, implement light toggling

+++ app/src/main/java/fr/filau/polyhome/housemanagement/data/HouseManagementDataDevice.kt
@@ -4,7 +4,7 @@ data class HouseManagementDataDevice(
     val id: String,
     val type: String,
     val availableCommands: ArrayList<String>,
-    val opening: Int?,
+    val opening: Float?,
     val openingMode: Int?,
-    val power: Int?
+    val power: Float?
 )
